Ending work with controller ordinarytask

+++ UltWolfScheduleAPI/Controllers/OrdinaryTasksController.cs
@@ -2,6 +2,7 @@
 using System.Collections.Generic;
 using System.Linq;
 using System.Threading.Tasks;
+using Microsoft.AspNetCore.Authorization;
 using Microsoft.AspNetCore.Http;
 using Microsoft.AspNetCore.Mvc;
 using Microsoft.EntityFrameworkCore;
@@ -10,6 +11,7 @@ using UltWolfScheduleAPI.Models.Context;
 
 namespace UltWolfScheduleAPI.Controllers
 {
+    [Authorize]
     [Route("api/[controller]")]
     [ApiController]
     public class OrdinaryTasksController : ControllerBase
@@ -21,13 +23,17 @@ namespace UltWolfScheduleAPI.Controllers
             _context = context;
         }
 
-        // GET: api/OrdinaryTasks
+        // GET: api/OrdinaryTasks 
         [HttpGet]
-        public IEnumerable<OrdinaryTask> GetOrdTasks()
+        public IEnumerable<OrdinaryTask> GetOrdTasksForToday()
         {
-            return _context.OrdTasks;
+            return _context.OrdTasks.Where(m=>m.User.Id.ToString() == User.Identity.Name).Where(m=>m.Date.Date==DateTime.Now.Date);
         }
-
+        [HttpGet("all")]
+        public IEnumerable<OrdinaryTask> GetallOrdTasksy()
+        {
+            return _context.OrdTasks.Where(m => m.User.Id.ToString() == User.Identity.Name);
+        }
         // GET: api/OrdinaryTasks/5
         [HttpGet("{id}")]
         public async Task<IActionResult> GetOrdinaryTask([FromRoute] int id)
@@ -35,15 +41,12 @@ namespace UltWolfScheduleAPI.Controllers
             if (!ModelState.IsValid)
             {
                 return BadRequest(ModelState);
-            }
-
-            var ordinaryTask = await _context.OrdTasks.FindAsync(id);
-
+            } 
+            var ordinaryTask = await _context.OrdTasks.FindAsync(id); 
             if (ordinaryTask == null)
             {
                 return NotFound();
-            }
-
+            } 
             return Ok(ordinaryTask);
         }
 
@@ -59,6 +62,11 @@ namespace UltWolfScheduleAPI.Controllers
             if (id != ordinaryTask.Id)
             {
                 return BadRequest();
+            }
+            if (ordinaryTask.UserId.ToString() != User.Identity.Name)
+            {
+                ModelState.AddModelError("error", "Sorry you don`t have enought permision for that");
+                return BadRequest(ModelState);
             }
 
             _context.Entry(ordinaryTask).State = EntityState.Modified;
@@ -90,14 +98,13 @@ namespace UltWolfScheduleAPI.Controllers
             {
                 return BadRequest(ModelState);
             }
-
+            ordinaryTask.UserId = int.Parse(User.Identity.Name);
             _context.OrdTasks.Add(ordinaryTask);
             await _context.SaveChangesAsync();
 
             return CreatedAtAction("GetOrdinaryTask", new { id = ordinaryTask.Id }, ordinaryTask);
         }
-
-        // DELETE: api/OrdinaryTasks/5
+         
         [HttpDelete("{id}")]
         public async Task<IActionResult> DeleteOrdinaryTask([FromRoute] int id)
         {
@@ -111,7 +118,11 @@ namespace UltWolfScheduleAPI.Controllers
             {
                 return NotFound();
             }
-
+            if(ordinaryTask.UserId.ToString() != User.Identity.Name)
+            {
+                ModelState.AddModelError("error", "Sorry you don`t have enought permision for that");
+                return BadRequest(ModelState);
+            }
             _context.OrdTasks.Remove(ordinaryTask);
             await _context.SaveChangesAsync();
